requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-telegram-bot==21.1.1
python-telegram-bot-calendar==1.0.5
SQLAlchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.4
pydantic==2.7.2
pydantic-settings==2.3.1
python-dotenv==1.0.1
